requirements.txt
absl-py==2.1.0
ai-edge-litert==1.1.1
ai-edge-litert-nightly==1.1.2.dev20250303
ai-edge-quantizer-nightly==0.0.1.dev20250208
ai-edge-torch==0.3.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
idna==3.10
immutabledict==4.2.1
iniconfig==2.0.0
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.5
keras-nightly==3.9.0.dev2025030403
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scipy==1.15.2
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.19.0a20250218
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_nightly==2.19.0.dev20250207
tomli==2.2.1
torch==2.6.0
torch_xla2==0.0.1.dev202412041639
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
